+++ zcash_client_backend/src/data_api/ll/wallet.rs
@@ -123,6 +123,8 @@ where
     Ok(())
 }
 
+#[derive(Debug)]
+#[non_exhaustive]
 pub enum PutBlocksError<SE, TE> {
     /// Returned if a provided block sequence has gaps.
     NonSequentialBlocks {