+++ zcash_client_sqlite/src/wallet/init.rs
@@ -30,6 +30,7 @@ const MIGRATIONS_TABLE: &str = "schemer_migrations";
 
 /// Errors that can occur when applying migrations to the wallet database.
 #[derive(Debug)]
+#[non_exhaustive]
 pub enum WalletMigrationError {
     /// A feature required by the wallet database is not supported by the version of
     /// SQLite that the migration is running against.
@@ -205,7 +206,8 @@ fn sqlite_client_error_to_wallet_migration_error(e: SqliteClientError) -> Wallet
         | SqliteClientError::KeyDerivationError(_)
         | SqliteClientError::Zip32AccountIndexOutOfRange
         | SqliteClientError::AccountCollision(_)
-        | SqliteClientError::CacheMiss(_) => {
+        | SqliteClientError::CacheMiss(_)
+        | SqliteClientError::BackendError(_) => {
             unreachable!("we only call WalletRead methods; mutations can't occur")
         }
         #[cfg(feature = "transparent-inputs")]