+++ zcash_keys/src/keys/transparent/gap_limits.rs
@@ -246,6 +246,7 @@ pub fn generate_address_list(
 }
 
 /// Errors that can occur when generating transparent gap addresses.
+#[derive(Debug)]
 pub enum GapAddressesError<SE> {
     /// An error occurred in the underlying wallet storage backend.
     Storage(SE),